requirements.txt
Django>=4.2
mssql-django>=1.3
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
pytz
python-dotenv
